172 edit and delete vehicle

+++ lib/actions/vehicles.tsx
@@ -1,17 +1,18 @@
-'use server';
+"use server";
 
-import Vehicle from '@/models/Vehicle';
-import { z } from 'zod';
-import { revalidatePath } from 'next/cache';
-import { ObjectId } from 'mongodb';
-import mongoose from 'mongoose';
+import Vehicle from "@/models/Vehicle";
+import { z } from "zod";
+import { revalidatePath } from "next/cache";
+import { ObjectId } from "mongodb";
+import mongoose from "mongoose";
+import { getFutureJobsByVehicleId } from "./jobs";
 
 const fieldFriendlyNames: Record<string, string> = {
-  licencePlate: 'Licence Plate',
-  gpsApi: 'GPS API',
-  make: 'Make',
-  model: 'Model',
-  status: 'Status',
+  licencePlate: "Licence Plate",
+  gpsApi: "GPS API",
+  make: "Make",
+  model: "Model",
+  status: "Status",
 };
 
 const addVehicle = async (vehicle: {
@@ -26,7 +27,7 @@ const addVehicle = async (vehicle: {
     gpsApi: z.string().min(1),
     make: z.string().min(1),
     model: z.string().min(1),
-    status: z.enum(['Draft', 'Active', 'Disabled']),
+    status: z.enum(["Draft", "Active", "Disabled"]),
   });
 
   const response = vehicleSchema.safeParse({
@@ -38,28 +39,34 @@ const addVehicle = async (vehicle: {
   });
 
   if (!response.success) {
-    return { message: '', errors: response.error.flatten().fieldErrors };
+    return { message: "", errors: response.error.flatten().fieldErrors };
   }
 
   try {
     const newVehicle = new Vehicle(response.data);
     await newVehicle.save();
 
-    return { message: 'Vehicle added successfully' };
+    return { message: "Vehicle added successfully" };
   } catch (error: unknown) {
     if (error instanceof mongoose.Error.ValidationError && error.errors) {
       // Mongoose validation errors (including unique-validator errors)
-      const mongooseErrors = Object.keys(error.errors).reduce((acc, key) => {
-        const friendlyKey = fieldFriendlyNames[key] || key; // Map to friendly name if available
-        const errorMessage = error.errors[key].message.replace(key, friendlyKey); // Replace field name in the message
-        acc[friendlyKey] = [errorMessage]; // Structure as an array to match Zod format
-        return acc;
-      }, {} as Record<string, string[]>);
+      const mongooseErrors = Object.keys(error.errors).reduce(
+        (acc, key) => {
+          const friendlyKey = fieldFriendlyNames[key] || key; // Map to friendly name if available
+          const errorMessage = error.errors[key].message.replace(
+            key,
+            friendlyKey
+          ); // Replace field name in the message
+          acc[friendlyKey] = [errorMessage]; // Structure as an array to match Zod format
+          return acc;
+        },
+        {} as Record<string, string[]>
+      );
 
-      return { message: 'Validation error', errors: mongooseErrors };
+      return { message: "Validation error", errors: mongooseErrors };
     }
 
-    return { message: 'An unexpected error occurred' };
+    return { message: "An unexpected error occurred" };
   }
 };
 
@@ -77,7 +84,7 @@ const updateVehicle = async (vehicle: {
     gpsApi: z.string().min(1),
     make: z.string(),
     model: z.string(),
-    status: z.enum(['Draft', 'Active', 'Disabled']),
+    status: z.enum(["Draft", "Active", "Disabled"]),
   });
 
   const response = vehicleSchema.safeParse({
@@ -90,39 +97,72 @@ const updateVehicle = async (vehicle: {
   });
 
   if (!response.success) {
-    return { message: '', errors: response.error.flatten().fieldErrors };
+    return { message: "", errors: response.error.flatten().fieldErrors };
   }
 
   const filter = { _id: new ObjectId(response.data._id) };
-  const update = { licencePlate: response.data.licencePlate, gpsApi: response.data.gpsApi, make: response.data.make, model: response.data.model, status: response.data.status };
-  const context = { runValidators: true, context: 'query' };
+  const update = {
+    licencePlate: response.data.licencePlate,
+    gpsApi: response.data.gpsApi,
+    make: response.data.make,
+    model: response.data.model,
+    status: response.data.status,
+  };
+  const context = { runValidators: true, context: "query" };
+
+  // if marking as inactive
+  // check for future jobs that use this vehicle
+  if (update.status === "Disabled") {
+    const futureJobs = await getFutureJobsByVehicleId(vehicle._id);
+    if (futureJobs.length > 0) {
+      return {
+        message: "Vehicle is being used in a future job.",
+        errors: {
+          status: [
+            "Jobs must be reassigned before vehicle can be marked inactive",
+          ],
+        },
+      };
+    }
+  }
 
   try {
     await Vehicle.findOneAndUpdate(filter, update, context);
-    revalidatePath('/staff/vehicles');
-    return { message: 'Vehicle updated successfully' };
+    revalidatePath("/staff/vehicles");
+    return { message: "Vehicle updated successfully" };
   } catch (error: unknown) {
     if (error instanceof mongoose.Error.ValidationError && error.errors) {
       // Mongoose validation errors (including unique-validator errors)
-      const mongooseErrors = Object.keys(error.errors).reduce((acc, key) => {
-        const friendlyKey = fieldFriendlyNames[key] || key; // Map to friendly name if available
-        const errorMessage = error.errors[key].message.replace(key, friendlyKey); // Replace field name in the message
-        acc[friendlyKey] = [errorMessage]; // Structure as an array to match Zod format
-        return acc;
-      }, {} as Record<string, string[]>);
+      const mongooseErrors = Object.keys(error.errors).reduce(
+        (acc, key) => {
+          const friendlyKey = fieldFriendlyNames[key] || key; // Map to friendly name if available
+          const errorMessage = error.errors[key].message.replace(
+            key,
+            friendlyKey
+          ); // Replace field name in the message
+          acc[friendlyKey] = [errorMessage]; // Structure as an array to match Zod format
+          return acc;
+        },
+        {} as Record<string, string[]>
+      );
 
-      return { message: 'Validation error', errors: mongooseErrors };
+      return { message: "Validation error", errors: mongooseErrors };
     }
 
     // Handle other types of errors (optional)
-    return { message: 'An unexpected error occurred' };
+    return { message: "An unexpected error occurred" };
   }
-
 };
 
 const deleteVehicle = async (vehicleId: string) => {
+  const futureJobs = await getFutureJobsByVehicleId(vehicleId);
+  if (futureJobs.length > 0) {
+    throw new Error(
+      "Vehicle is being used in a future job. Job must be reassigned before vehicle can be deleted."
+    );
+  }
   await Vehicle.findByIdAndDelete(vehicleId);
-  revalidatePath('/staff/vehicles');
+  revalidatePath("/staff/vehicles");
 };
 
 const getVehicle = async (vehicleId: string) => {
